Click mouse-wheel to optimize text for file-name
Sort the titles by time. 
Simplify time format, removing "S" at the end, remove "0H" if the time is smaller than 1 hour.
Remove the blanks adjacent to ";".
If the text is longer than 230 (max length of file-name), split the text into 230 and the longer part. The buttons are based on the whole text.